# api/openapi.yaml
openapi: 3.0.3
info:
  title: API ImmoConnect
  description: >
    API RESTful de la plateforme ImmoConnect (agence immobilière) :
    consultation des biens, favoris, messagerie, prise de rendez-vous de visite
    avec créneaux premium payants (Stripe), blog et volet Open Data.
    Documentation générée avec springdoc-openapi et servie par Swagger UI.
  version: 1.0.0
  contact:
    name: ImmoConnect — support API
    email: api@immoconnect.be
  license:
    name: Données Open Data sous licence CC BY 4.0
    url: https://creativecommons.org/licenses/by/4.0/deed.fr
servers:
  - url: https://www.immoconnect.be/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Développement local
...
security:
  - bearerAuth: []
paths:
...
  /signalements:
    post:
      tags: [Signalements]
      summary: Signaler un contenu
      description: Un message reçu, une annonce ou un article (règlement européen sur les services numériques). Un seul signalement ouvert par contenu et par personne ; l'administrateur tranche et motive sa décision.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignalementRequete' }
      responses:
        '201':
          description: Signalement enregistré
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Signalement' }
        '403': { $ref: '#/components/responses/ErreurDroits' }
        '404': { $ref: '#/components/responses/ErreurIntrouvable' }
        '422': { $ref: '#/components/responses/ErreurValidation' }
...
  /admin/utilisateurs:
    get:
      tags: [Administration]
      summary: Comptes de la plateforme (gestionnaire)
      parameters:
        - { name: role, in: query, schema: { type: string, enum: [membre, agent, admin] } }
        - { name: recherche, in: query, description: Nom, prénom ou e-mail, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: taille, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200':
          description: Page de comptes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageComptes' }
        '403': { $ref: '#/components/responses/ErreurDroits' }
  /admin/agents:
    post:
      tags: [Administration]
      summary: Ouvrir le compte d'un agent (gestionnaire)
      description: Le matricule est attribué par le serveur ; l'agent change ensuite son mot de passe provisoire.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nom, prenom, email, telephonePro, motDePasse]
              properties:
                nom: { type: string, maxLength: 80 }
                prenom: { type: string, maxLength: 80 }
                email: { type: string, format: email }
                telephonePro: { type: string, example: '+32 470 12 34 56' }
                langue: { type: string, enum: [fr, nl, en], default: fr }
                motDePasse: { type: string, format: password, minLength: 8 }
      responses:
        '201':
          description: Compte créé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Compte' }
        '409': { $ref: '#/components/responses/ErreurConflit' }
        '422': { $ref: '#/components/responses/ErreurValidation' }
  /admin/utilisateurs/{id}/desactiver:
    patch:
      tags: [Administration]
      summary: Désactiver un compte (gestionnaire)
      description: >
        Le compte ne peut plus se connecter ; son historique est conservé. Nul ne se désactive
        lui-même, et seul un super-administrateur agit sur le compte d'un administrateur.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Compte désactivé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Compte' }
        '403': { $ref: '#/components/responses/ErreurDroits' }
  /admin/utilisateurs/{id}/activer:
    patch:
      tags: [Administration]
      summary: Réactiver un compte (gestionnaire)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Compte réactivé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Compte' }
  /admin/journal:
    get:
      tags: [Administration]
      summary: Journal d'audit (gestionnaire)
      description: Du plus récent au plus ancien. Le journal se consulte ; aucune opération ne le modifie.
      parameters:
        - { name: action, in: query, schema: { type: string, example: connexion } }
        - { name: utilisateurId, in: query, schema: { type: integer } }
        - { name: du, in: query, schema: { type: string, format: date } }
        - { name: au, in: query, schema: { type: string, format: date } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: taille, in: query, schema: { type: integer, default: 50, maximum: 100 } }
      responses:
        '200':
          description: Page de traces
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageTraces' }
...
  /admin/categories:
    post:
      tags: [Administration]
      summary: Créer une catégorie de biens (éditeur)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategorieRequete' }
      responses:
        '201':
          description: Catégorie créée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Categorie' }
        '422': { $ref: '#/components/responses/ErreurValidation' }
  /admin/categories/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    put:
      tags: [Administration]
      summary: Modifier une catégorie de biens (éditeur)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategorieRequete' }
      responses:
        '200':
          description: Catégorie mise à jour
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Categorie' }
    delete:
      tags: [Administration]
      summary: Supprimer une catégorie de biens (éditeur)
      description: Refusé (409) tant que des biens portent cette catégorie.
      responses:
        '204': { description: Catégorie supprimée }
        '409': { $ref: '#/components/responses/ErreurConflit' }
...
  /admin/traductions/{cle}:
    parameters:
      - { name: cle, in: path, required: true, schema: { type: string, example: accueil.titre } }
    put:
      tags: [Administration]
      summary: Créer ou modifier un texte du site (éditeur)
      description: Seules les langues transmises sont modifiées. Effet immédiat sur GET /traductions/{codeLangue}.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [valeurs]
              properties:
                valeurs:
                  type: object
                  additionalProperties: { type: string, maxLength: 2000 }
                  example: { fr: Trouvez le bien qui vous ressemble, nl: Vind het pand dat bij u past }
      responses:
        '200':
          description: Texte enregistré
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TraductionLigne' }
        '422': { $ref: '#/components/responses/ErreurValidation' }
    delete:
      tags: [Administration]
      summary: Supprimer un texte du site dans toutes les langues (éditeur)
      responses:
        '204': { description: Texte supprimé }
        '404': { $ref: '#/components/responses/ErreurIntrouvable' }
  /admin/cles-api:
    get:
      tags: [Administration]
      summary: Clés API délivrées (gestionnaire)
      description: La clé elle-même n'est jamais renvoyée — le serveur n'en conserve que l'empreinte SHA-256.
      responses:
        '200':
          description: Clés, de la plus récente à la plus ancienne
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CleApi' }
    post:
      tags: [Administration]
      summary: Générer une clé API (gestionnaire)
      description: La clé figure dans cette réponse et nulle part ailleurs — elle ne pourra plus être affichée.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [libelle]
              properties:
                libelle: { type: string, maxLength: 80, example: Observatoire des loyers }
      responses:
        '201':
          description: Clé créée, valeur comprise
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CleApi' }
  /admin/cles-api/{id}/revoquer:
    patch:
      tags: [Administration]
      summary: Révoquer une clé API (gestionnaire)
      description: Effet immédiat et définitif (règle RA12).
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Clé révoquée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CleApi' }
  /admin/alertes:
    get:
      tags: [Administration]
      summary: Alertes de sécurité (gestionnaire)
      description: >
        Détection d'intrusion : rafale d'échecs de connexion, énumération d'identifiants (l'adresse est alors
        bannie de la connexion pendant quinze minutes), clé API révoquée présentée. Chaque alerte est aussi
        envoyée par e-mail aux super-administrateurs.
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: taille, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200':
          description: De la plus récente à la plus ancienne
          content:
            application/json:
              schema:
                type: object
                properties:
                  contenu: { type: array, items: { $ref: '#/components/schemas/AlerteSecurite' } }
                  page: { type: integer }
                  taille: { type: integer }
                  totalElements: { type: integer }
                  totalPages: { type: integer }
        '403': { $ref: '#/components/responses/ErreurDroits' }
  /admin/parametres:
    get:
      tags: [Administration]
      summary: Paramètres du site (gestionnaire, cas A5)
      responses:
        '200':
          description: Nom, slogan, coordonnées et horaires de l'agence, langues actives
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Parametre' }
        '403': { $ref: '#/components/responses/ErreurDroits' }
    put:
      tags: [Administration]
      summary: Modifier des paramètres (gestionnaire)
      description: Valeurs par clé ; une clé inconnue ou une valeur invalide est refusée. Le français reste toujours actif. Chaque changement est journalisé.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [valeurs]
              properties:
                valeurs:
                  type: object
                  additionalProperties: { type: string }
                  example: { agence.nom: ImmoConnect Bruxelles, langues.actives: 'fr,nl' }
      responses:
        '200':
          description: Paramètres après modification
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Parametre' }
        '422': { $ref: '#/components/responses/ErreurValidation' }
  /admin/signalements:
    get:
      tags: [Administration]
      summary: Signalements de contenus (gestionnaire)
      parameters:
        - { name: statut, in: query, schema: { type: string, enum: [ouvert, retire, conserve] } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: taille, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200':
          description: Du plus récent au plus ancien
          content:
            application/json:
              schema:
                type: object
                properties:
                  contenu: { type: array, items: { $ref: '#/components/schemas/Signalement' } }
                  page: { type: integer }
                  taille: { type: integer }
                  totalElements: { type: integer }
                  totalPages: { type: integer }
        '403': { $ref: '#/components/responses/ErreurDroits' }
...
  /admin/signalements/{id}:
    patch:
      tags: [Administration]
      summary: Trancher un signalement (gestionnaire)
      description: « retire » vide le message, met l'annonce hors ligne ou archive l'article ; « conserve » le laisse en place. La décision est motivée, définitive et journalisée (RA13).
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [statut, decision]
              properties:
                statut: { type: string, enum: [retire, conserve] }
                decision: { type: string, maxLength: 500, example: Annonce dupliquée, retirée en attendant les explications de l'agent. }
      responses:
        '200':
          description: Signalement tranché
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Signalement' }
        '404': { $ref: '#/components/responses/ErreurIntrouvable' }
        '422': { $ref: '#/components/responses/ErreurValidation' }
  /traductions/{codeLangue}:
    get:
      tags: [Traductions]
      summary: Dictionnaire de traduction de l'interface (fr, nl, en)
      security: []
      parameters:
        - { name: codeLangue, in: path, required: true, schema: { type: string, enum: [fr, nl, en] } }
      responses:
        '200':
          description: Paires clé/valeur
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { type: string }
              example: { nav.accueil: Accueil, rdv.bouton: Prendre rendez-vous }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Jeton JWT obtenu via POST /auth/login (en-tête Authorization: Bearer <jeton>)"
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API délivrée par l'administrateur (table cle_api) — révocable (règle RA12)
  responses:
    ErreurValidation:
      description: Données invalides (règles du dictionnaire de données)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Erreur' }
...
    ErreurDroits:
      description: Rôle insuffisant pour cette opération
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Erreur' }
...
    ErreurIntrouvable:
      description: Ressource inexistante
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Erreur' }
    ErreurConflit:
      description: Conflit d'état (créneau déjà réservé — A2, transition interdite — RA1)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Erreur' }
...
  schemas:
...
    AlerteSecurite:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, enum: [rafale_echecs, enumeration, cle_revoquee] }
        ip: { type: string, example: 203.0.113.9 }
        detail: { type: string, example: 7 comptes différents essayés en 10 min }
        creeLe: { type: string, format: date-time }
    Parametre:
      type: object
      properties:
        cle: { type: string, enum: [agence.nom, agence.slogan, agence.adresse, agence.telephone, agence.email, agence.horaires, langues.actives] }
        valeur: { type: string, maxLength: 1000 }
        modifieLe: { type: string, format: date-time }
        modifiePar: { type: string, nullable: true, example: Lotte Goossens }
    SignalementRequete:
      type: object
      required: [typeContenu, contenuId, motif, description]
      properties:
        typeContenu: { type: string, enum: [message, bien, article] }
        contenuId: { type: integer }
        motif: { type: string, enum: [illicite, arnaque, indesirable, autre] }
        description: { type: string, maxLength: 1000 }
    Signalement:
      type: object
      properties:
        id: { type: integer }
        typeContenu: { type: string, enum: [message, bien, article] }
        contenuId: { type: integer }
        apercu: { type: string, nullable: true, description: Début du contenu visé }
        motif: { type: string, enum: [illicite, arnaque, indesirable, autre] }
        description: { type: string }
        statut: { type: string, enum: [ouvert, retire, conserve] }
        creeLe: { type: string, format: date-time }
        auteur: { type: string }
        auteurRole: { type: string, enum: [membre, agent, admin] }
        traiteLe: { type: string, format: date-time, nullable: true }
        traitePar: { type: string, nullable: true }
        decision: { type: string, nullable: true }
...
    Categorie:
      type: object
      properties:
        id: { type: integer, example: 2 }
        nom: { type: string, example: Appartement }
        description: { type: string, nullable: true }
    CategorieRequete:
      type: object
      required: [nom]
      properties:
        nom: { type: string, maxLength: 60 }
        description: { type: string, maxLength: 1000 }
...
    Compte:
      type: object
      description: Compte vu par l'administrateur — le mot de passe, même haché, n'est jamais exposé
      properties:
        id: { type: integer, example: 101 }
        prenom: { type: string, example: Sarah }
        nom: { type: string, example: Dubois }
        email: { type: string, format: email }
        role: { type: string, enum: [membre, agent, admin] }
        langue: { type: string, example: fr }
        actif: { type: boolean }
        dateInscription: { type: string, format: date }
        matricule: { type: string, nullable: true, example: AG-2023-001 }
        niveauAcces: { type: integer, nullable: true, example: 2 }
    PageComptes:
      type: object
      properties:
        contenu: { type: array, items: { $ref: '#/components/schemas/Compte' } }
        page: { type: integer }
        taille: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
...
    PageTraces:
      type: object
      properties:
        contenu: { type: array, items: { $ref: '#/components/schemas/TraceAudit' } }
        page: { type: integer }
        taille: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
    CleApi:
      type: object
      properties:
        id: { type: integer }
        libelle: { type: string, nullable: true, example: Observatoire des loyers }
        creeLe: { type: string, format: date }
        active: { type: boolean }
        derniereUtilisation: { type: string, format: date-time, nullable: true }
        creeePar: { type: string, example: Lotte Goossens }
        cle:
          type: string
          nullable: true
          example: ic_a762098493b17f323023fa5819374854997fddaa7364a62c5f83012c187f680a
          description: Renseignée uniquement dans la réponse à la création
    TraductionLigne:
      type: object
      properties:
        cle: { type: string, example: accueil.titre }
        valeurs:
          type: object
          additionalProperties: { type: string }
          example: { fr: Trouvez le bien qui vous ressemble, nl: Vind het pand dat bij u past, en: Find the property that suits you }
...